controls separated but not working
